Removed 2D MPFAO code.
All functions that indicate they were intended for 3D calculations only
(e.g. foo_3DMesh) have been renamed to exclude that indication.
